+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for ordering search results by the indexed templates settings. Sort setting needs to be `_indexed_templates` in order for this feature to kick in.
+
 ## [0.28.0] - 2021-02-14
 
 ### Added